Edit code to utilize a rest parameter and .includes
This change in code allows the function to pass all required tests. The rest parameter allows for multiple parameters to be passed for this parameter. The .includes array method, in this case, allows us to NOT include items in the parameter when filtering. This way, the filter will exclude whatever is listed in the parameter.